+++ src/net/tmwa/charserverhandler.cpp
@@ -157,9 +157,9 @@ void CharServerHandler::readPlayerData(Net::MessageIn &msg,
     data.mAttributes[PlayerInfo::MAX_MP] = msg.readInt16();
 
     msg.readInt16();                           // speed
-    const int race = msg.readInt16();          // class (used for race)
+    const uint16_t race = msg.readInt16();          // class (used for race)
     const int hairStyle = msg.readInt8();
-    const int look = msg.readInt8();
+    const uint8_t look = msg.readInt8();
     tempPlayer->setSubtype(race, look);
     const uint16_t weapon = msg.readInt16();  // unused on server. need use?
     tempPlayer->setSprite(SPRITE_WEAPON, weapon, "", 1, true);
